Add rust-toolchain, ie. pins Rust version
The rust-toolchain is just copied from the rustgpu repo, as it seems the
most fussy about which Rust it uses.

I also changed the references from mitchmindtree's old fork of rustgpu
to the native Embark repo, and pinned to the lastest commit. Might have
been better to pin to the latest tag, but I'd already had a day full of
compiling (I seem to have to compile for rust-analyzer as well??).
Anyway it all works now, and everything that can be pinned is pinned.